requirements.txt
requests==2.31.0
numpy==1.26.2
faiss-cpu==1.7.4
tqdm==4.66.1
python-dotenv==1.0.0
aiohttp==3.9.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
APScheduler==3.10.4
